Create Assignment Task-2 (Code)
This is a simple Python program that asks a person to type in their first and last names. Then, it joins these two names together to make a full name and shows a friendly message using that name. It is an easy way to learn how computers can take information from people and respond with a personalized greeting.